+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "2.5D maze".
 - Tests for "Minimal number of swaps".
 - Tests for "Cylinders".
+- Tests for "Killer sudoku solver".
 
 ## [1.16.0] - 2022-12-31
 ### Added